Refactor layout code

+++ src/tree_store/page_store/region.rs
@@ -19,15 +19,15 @@ impl<'a> RegionHeaderMutator<'a> {
         Self { mem: data }
     }
 
-    pub(crate) fn initialize(&mut self, num_pages: usize, max_page_capacity: usize) {
+    pub(crate) fn initialize(&mut self, num_pages: u32, max_page_capacity: u32) {
         self.mem[0] = REGION_FORMAT_VERSION;
-        let allocator_len = BuddyAllocatorMut::required_space(max_page_capacity);
+        let allocator_len = BuddyAllocatorMut::required_space(max_page_capacity as usize);
         self.mem[ALLOCATOR_LENGTH_OFFSET..(ALLOCATOR_LENGTH_OFFSET + size_of::<u32>())]
             .copy_from_slice(&(allocator_len as u32).to_le_bytes());
         BuddyAllocatorMut::init_new(
             &mut self.mem[ALLOCATOR_OFFSET..(ALLOCATOR_OFFSET + allocator_len)],
-            num_pages,
-            max_page_capacity,
+            num_pages as usize,
+            max_page_capacity as usize,
         );
     }
